added reset password component, change password in myaccount, check token expired

+++ area_Amministrativa/src/app/modale-delete/modale-delete.component.ts
@@ -1,6 +1,8 @@
 import {Component, Inject} from '@angular/core';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
-import {HttpProviderService} from "../service/http-provider.service";
+import { HttpProviderService } from "../service/http-provider.service";
+import { AuthService } from "../service/auth.service";
+import { ToastrService } from 'ngx-toastr';
 
 @Component({
   selector: 'app-modale-delete',
@@ -8,27 +10,36 @@ import {HttpProviderService} from "../service/http-provider.service";
   styleUrls: ['./modale-delete.component.css']
 })
 export class ModaleDeleteComponent {
-  constructor(private ref: MatDialogRef<ModaleDeleteComponent>,
+  constructor(public auth: AuthService, private toastr: ToastrService, private ref: MatDialogRef<ModaleDeleteComponent>,
     @Inject(MAT_DIALOG_DATA) public data: { Id: string, ClassForm : string },
                private httpApi: HttpProviderService) {}
 
   // delete for call DELETE API
   confirmForm(): void {
-    if (this.data.ClassForm == "User") {
-      //this.httpApi.deleteUser(this.data.Id).subscribe()
-    }
-    else if (this.data.ClassForm == "People") {
-      this.httpApi.deletePerson(this.data.Id).subscribe()
-    }
-    else if (this.data.ClassForm == "Organization") {
-      //this.httpApi.deleteOrg(this.data.Id).subscribe()
-    }
+    if (this.auth.isAuthenticated()) {
+      if (this.data.ClassForm == "User") {
+        //this.httpApi.deleteUser(this.data.Id).subscribe()
+      }
+      else if (this.data.ClassForm == "People") {
+        this.httpApi.deletePerson(this.data.Id).subscribe()
+      }
+      else if (this.data.ClassForm == "Organization") {
+        //this.httpApi.deleteOrg(this.data.Id).subscribe()
+      }
 
-    console.log("delete: " + this.data.Id)
-    console.log("type: " + this.data.ClassForm)
+      console.log("delete: " + this.data.Id)
+      console.log("type: " + this.data.ClassForm)
 
-    this.ref.close()
-    window.location.reload();
+      this.ref.close()
+      window.location.reload();
+    }
+    else {
+      this.toastr.error("Token is expired", "Error")
+      setTimeout(() => {
+        window.location.reload();
+      }, 500)
+      
+    }
   }
 
   closepopup() {
